+++ Installomator.sh
@@ -20,8 +20,8 @@
 #set -o xtrace # outputting every command of the script
 #set -x # Debug
 
-VERSION='0.4.13' # This version branched by Søren Theilgaard
-VERSIONDATE='2020-12-21'
+VERSION='0.4.14' # This version branched by Søren Theilgaard
+VERSIONDATE='2020-12-23'
 VERSIONBRANCH='Søren Theilgaard'
 
 export PATH=/usr/bin:/bin:/usr/sbin:/sbin